+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - add support for STEP files format (needs opencascade lib) #1222
 - fix nan case in scalarbar3d()
 - add `.rename()` method to set any object name.
+- fix bug #1230 in `line.find_index_at_position()` by @natabma
 
 
 
@@ -31,6 +32,8 @@ Changes that will break existing code and need active thinking and some work to 
 ## New/Revised Examples
 ```
 examples/pyplot/plot_fxy0.py
+
+tests/issues/issue_1230.py
 ```
 
 ## To Do